- инициализация instance

+++ Project/FPS/Renderer.cpp
@@ -6,9 +6,10 @@
 namespace
 {
 	bool UseValidationLayers{ false };
-
 	VkInstance Instance{ nullptr };
 	VkDebugUtilsMessengerEXT DebugMessenger{ nullptr };
+
+
 	VkPhysicalDevice ChosenGPU{ nullptr };
 	VkDevice Device{ nullptr };
 	VkSurfaceKHR Surface{ nullptr };
@@ -27,7 +28,6 @@ bool initVulkan()
 		return false;
 	}
 
-	// current version vulkan
 	{
 		uint32_t version{ 0 };
 		vkEnumerateInstanceVersion(&version);
@@ -43,25 +43,37 @@ bool initVulkan()
 		// TODO: делать проверку на версию
 	}
 
-	vkb::InstanceBuilder builder;
+	vkb::InstanceBuilder instanceBuilder;
 
-	auto inst_ret = builder.set_app_name("FPS Game")
+	auto instanceRet = instanceBuilder
 		.request_validation_layers(UseValidationLayers)
 		.use_default_debug_messenger()
 		.require_api_version(1, 3, 0)
 		.build();
+	if (!instanceRet)
+	{
+		Fatal(instanceRet.error().message());
+		return false;
+	}
 
-	vkb::Instance vkbInst = inst_ret.value();
+	vkb::Instance vkbInstance = instanceRet.value();
 
-	Instance = vkbInst.instance;
-	DebugMessenger = vkbInst.debug_messenger;
+	Instance = vkbInstance.instance;
+	DebugMessenger = vkbInstance.debug_messenger;
 
 	volkLoadInstanceOnly(Instance);
 
 	result = glfwCreateWindowSurface(Instance, Window::GetWindow(), nullptr, &Surface);
 	if (result != VK_SUCCESS)
 	{
-		Fatal("glfwCreateWindowSurface failed!");
+		const char* errorMsg = nullptr;
+		int ret = glfwGetError(&errorMsg);
+		if (ret != 0)
+		{
+			std::string text = std::to_string(ret) + " ";
+			if (errorMsg != nullptr) text += std::string(errorMsg);
+			Fatal(text);
+		}
 		return false;
 	}
 
@@ -78,23 +90,31 @@ bool initVulkan()
 	// vulkan 1.1 features
 	VkPhysicalDeviceVulkan11Features features11{ .sType = VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_VULKAN_1_1_FEATURES };
 
-	// use vkbootstrap to select a gpu.
-	vkb::PhysicalDeviceSelector selector{ vkbInst };
-	vkb::PhysicalDevice physicalDevice = selector
+	vkb::PhysicalDeviceSelector physicalDeviceSelector{ vkbInstance };
+	auto physicalDeviceRet = physicalDeviceSelector
 		.set_minimum_version(1, 3)
 		.set_required_features_13(features13)
 		.set_required_features_12(features12)
 		.set_required_features_11(features11)
 		.set_surface(Surface)
-		.select()
-		.value();
+		.select();
+	if (!physicalDeviceRet) 
+	{
+		Fatal(physicalDeviceRet.error().message());
+		return false;
+	}
 
-	// create the final vulkan device
+	vkb::PhysicalDevice physicalDevice = physicalDeviceRet.value();
+
 	vkb::DeviceBuilder deviceBuilder{ physicalDevice };
+	auto deviceRet = deviceBuilder.build();
+	if (!deviceRet)
+	{
+		Fatal(deviceRet.error().message());
+		return false;
+	}
+	vkb::Device vkbDevice = deviceRet.value();
 
-	vkb::Device vkbDevice = deviceBuilder.build().value();
-
-	// Get the VkDevice handle used in the rest of a vulkan application
 	Device = vkbDevice.device;
 	ChosenGPU = physicalDevice.physical_device;
 
@@ -138,6 +158,11 @@ bool initVulkan()
 	return true;
 }
 
+bool createSwapchain()
+{
+	return true;
+}
+
 bool Renderer::Init()
 {
 #ifdef _DEBUG
@@ -145,6 +170,7 @@ bool Renderer::Init()
 #endif
 
 	if (!initVulkan()) return false;
+	if (!createSwapchain()) return false;
 
 	return true;
 }
@@ -153,14 +179,15 @@ void Renderer::Close()
 {
 	if (Device) vkDeviceWaitIdle(Device);
 
-	if (Allocator) vmaDestroyAllocator(Allocator);
 
-	if (Instance && Surface) vkDestroySurfaceKHR(Instance, Surface, nullptr);
+
+	if (Allocator) vmaDestroyAllocator(Allocator);
 	if (Device) vkDestroyDevice(Device, nullptr);
 
 	if (Instance)
 	{
-		vkb::destroy_debug_utils_messenger(Instance, DebugMessenger);
+		if (Surface) vkDestroySurfaceKHR(Instance, Surface, nullptr);
+		if (DebugMessenger) vkb::destroy_debug_utils_messenger(Instance, DebugMessenger);
 		vkDestroyInstance(Instance, nullptr);
 	}
